moved all call-related definitions to primitives (#2583)
